+++ app/components/PageSection/CardsLayout/QuadComponent.tsx
